warings from gcc on windows
[git-p4: depot-paths = "//depot/qt/main/": change = 161846]

+++ src/compat/other/q3process_win.cpp
@@ -538,7 +538,7 @@ void Q3Process::socketWrite( int )
 	    return;
 	}
 	d->stdinBufRead += written;
-	if ( d->stdinBufRead == d->stdinBuf.head()->size() ) {
+	if ( d->stdinBufRead == (DWORD)d->stdinBuf.head()->size() ) {
 	    d->stdinBufRead = 0;
 	    delete d->stdinBuf.dequeue();
 	    if ( wroteToStdinConnected && d->stdinBuf.isEmpty() )
